requirements.txt
#-i http://mirrors.aliyun.com/pypi/simple/ --trusted-host mirrors.aliyun.com
-i https://pypi.doubanio.com/simple/
python-gitlab
Flask
Flask_SQLAlchemy
Flask-Session
Flask-Caching
redis
smart_getenv
SQLAlchemy
Werkzeug
python-etcd
flask_mako
protobuf
grpcio
raven
PyYAML
more_itertools
gunicorn
pymysql
requests
six
celery[redis]
humanize
humanfriendly
boltons
blinker
wrapt
pyparsing
crontab
marshmallow
-e git+http://gitlab.ricebook.net/platform/mapi.git#egg=mapi
-e git+http://gitlab.ricebook.net/platform/erulb3.git#egg=erulb-py
